book_mgmt/requirements.txt
asgiref==3.5.2
Django==3.2
Pillow==9.2.0
psycopg2-binary==2.9.3
python-decouple==3.3
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
django-cors-headers==3.12.0
# dj-rest-auth==2.2.4
